fastapi/requirements.txt
# FastAPI and Server
fastapi==0.116.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic and Validation (already in requirements.txt but adding email support)
pydantic[email]==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Date and Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Type Checking
mypy==1.7.0
types-passlib==1.7.7.13

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Existing dependencies
annotated-types==0.7.0
anyio==4.10.0
fastapi-swagger-dark==0.0.7
idna==3.10
sniffio==1.3.1
starlette==0.47.3
typing-extensions==4.15.0
typing-inspection==0.4.1
